+++ src/main/java/chess/domain/board/KingInitializer.java
@@ -1,7 +1,7 @@
-package chess.domain.strategy.initializestategy;
+package chess.domain.board;
 
+import chess.domain.piece.King;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
 import chess.domain.piece.Team;
 import chess.domain.position.Position;
 
@@ -10,8 +10,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class QueenInitializer implements LocationInitializer {
-    private static final List<String> HORIZONTALS = Collections.singletonList("d");
+public class KingInitializer implements LocationInitializer {
+    private static final List<String> HORIZONTALS = Collections.singletonList("e");
     private static final List<String> VERTICALS_WHITE = Collections.singletonList("1");
     private static final List<String> VERTICALS_BLACK = Collections.singletonList("8");
 
@@ -19,8 +19,8 @@ public class QueenInitializer implements LocationInitializer {
     public Map<Position, Piece> initialize() {
         final Map<Position, Piece> chessBoard = new HashMap<>();
         for (String horizontal : HORIZONTALS) {
-            VERTICALS_BLACK.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new Queen(Team.BLACK)));
-            VERTICALS_WHITE.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new Queen(Team.WHITE)));
+            VERTICALS_BLACK.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new King(Team.BLACK)));
+            VERTICALS_WHITE.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new King(Team.WHITE)));
         }
         return chessBoard;
     }
