+++ src/main/java/chess/domain/board/LocationInitializer.java
@@ -1,4 +1,4 @@
-package chess.domain.strategy.initializestategy;
+package chess.domain.board;
 
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
