+++ src/main/java/chess/domain/board/PawnInitializer.java
@@ -1,4 +1,4 @@
-package chess.domain.strategy.initializestategy;
+package chess.domain.board;
 
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
