+++ src/main/java/chess/domain/board/RookInitializer.java
@@ -1,14 +1,14 @@
-package chess.domain.strategy.initializestategy;
+package chess.domain.board;
 
-import chess.domain.piece.Bishop;
 import chess.domain.piece.Piece;
+import chess.domain.piece.Rook;
 import chess.domain.piece.Team;
 import chess.domain.position.Position;
 
 import java.util.*;
 
-public class BishopInitializer implements LocationInitializer {
-    private static final List<String> HORIZONTALS = Arrays.asList("c", "f");
+public class RookInitializer implements LocationInitializer {
+    private static final List<String> HORIZONTALS = Arrays.asList("a", "h");
     private static final List<String> VERTICALS_WHITE = Collections.singletonList("1");
     private static final List<String> VERTICALS_BLACK = Collections.singletonList("8");
 
@@ -16,8 +16,8 @@ public class BishopInitializer implements LocationInitializer {
     public Map<Position, Piece> initialize() {
         final Map<Position, Piece> chessBoard = new HashMap<>();
         for (String horizontal : HORIZONTALS) {
-            VERTICALS_BLACK.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new Bishop(Team.BLACK)));
-            VERTICALS_WHITE.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new Bishop(Team.WHITE)));
+            VERTICALS_BLACK.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new Rook(Team.BLACK)));
+            VERTICALS_WHITE.forEach(vertical -> chessBoard.put(new Position(horizontal, vertical), new Rook(Team.WHITE)));
         }
         return chessBoard;
     }
